postgres_to_es/requirements.txt
backoff==2.2.1
elasticsearch==7.15.2
pydantic==1.10.2
psycopg2==2.9.5
python-dotenv==0.21.0